Move updateLocalStateOfRemoteConv to Action
